Continue styling the new design
