+++ assets/src/components/editor/Editor.tsx
@@ -41,10 +41,7 @@ export default (props: Props) => {
 
   return (
     <div
-      class={classNames(
-        'max-w-2xl mx-auto w-full h-auto grow pt-4 pb-20 px-8 bg-white rounded-md mt-4 shadow',
-        'print:shadow-none print:max-w-none print:p-0'
-      )}
+      class={classNames('max-w-2xl mx-auto w-full h-auto grow pb-20', 'print:max-w-none print:p-0')}
     >
       <div ref={editorRef} class="listnote mt-4 h-full" />
     </div>
